Bug/86/lesson model and pagination (#87)
* #86 - refactor lesson module

- add field ATTACHMETS to model
- fixed type for field CATEGORY in model on ResourceCategory
- add route /lessons/own for get own lessons
- add populate for fields CATEGORY and ATTACHMETS
- add sorting and filtering for collections

* Update lesson.controller.js

* Update lesson.routes.js

* Update index.js

* add get Categories with params

* fixed module Subject

- fixed field category in model
- fixed field name in model
- add functionality to get all subjects with params (name, category, skip, limit, sort)
- optimized code in  controller and service
- set correct method PATCH for update subject
- fixed errors in server terminal

---------

+++ src/modules/subject/subject.service.js
@@ -1,34 +1,49 @@
+const { createNotFoundError } = require('~/utils/errorsHelper')
 const Subject = require('./subject.model')
 
 const subjectService = {
-  getSubjects: async () => {
-    const items = await Subject.find().lean().exec()
-    const count = await Subject.countDocuments()
+  findMany: async (match, sort, skip, limit) => {
+    const count = await Subject.countDocuments(match)
+    const items = await Subject.find(match)
+      .collation({ locale: 'en', strength: 1 })
+      .populate([{ path: 'category', select: 'name' }])
+      .sort(sort)
+      .skip(skip)
+      .limit(limit)
+      .lean()
+      .exec()
+
     return { count, items }
   },
 
-  getSubjectById: async (id) => {
+  findOneById: async (id) => {
     return await Subject.findById(id).lean().exec()
   },
 
-  // Admin access
-
-  createSubject: async (data) => {
+  create: async (data) => {
     const subject = new Subject(data)
     await subject.save()
+
     return subject
   },
 
-  updateSubject: async (id, data) => {
-    const subject = await Subject.findByIdAndUpdate(id, data, { new: true })
-    return subject
+  update: async (id, data) => {
+    const subject = await subjectService.findOneById(id)
+    if (!subject) {
+      throw createNotFoundError()
+    }
+
+    return await Subject.findByIdAndUpdate(id, data, { new: true })
   },
 
-  deleteSubject: async (id) => {
-    const subject = await Subject.findByIdAndDelete(id)
-    return subject
-  },
+  remove: async (id) => {
+    const subject = await subjectService.findOneById(id)
+    if (!subject) {
+      throw createNotFoundError()
+    }
 
+    return Subject.findByIdAndDelete(id)
+  }
 }
 
 module.exports = subjectService
